requirements.txt
feedparser
beautifulsoup4
requests
google-genai
python-dotenv
truthbrush
snscrape
discord
cryptofeed>=2.4.0
websocket-client
aiosqlite
aiohttp
